+++ .projenrc.js
@@ -148,7 +148,7 @@ fetchAssemblies.exec(`node scripts/fetch-assemblies.js`);
 project.gitignore.exclude("public/data");
 
 // Proxy requests to awscdk.io for local testing
-project.package.addField("proxy", "https://d2ha9yenu2jur.cloudfront.net");
+project.package.addField("proxy", "https://d2wzz0rldl6xw0.cloudfront.net");
 
 // setup linting for create-react-app specific tools
 project.eslint.addRules({
