+++ package.json
@@ -119,6 +119,6 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "https://d2ha9yenu2jur.cloudfront.net",
+  "proxy": "https://d2wzz0rldl6xw0.cloudfront.net",
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
