--- src/api/docgen/view/initializer.ts
@@ -1,40 +1,0 @@
-import * as reflect from "jsii-reflect";
-import { Markdown } from "../render/markdown";
-import { Transpile, TranspiledCallable } from "../transpile/transpile";
-import { Parameter } from "./parameter";
-
-export class Initializer {
-  private readonly transpiled: TranspiledCallable;
-  private readonly parameters: Parameter[];
-  constructor(
-    private readonly transpile: Transpile,
-    initializer: reflect.Initializer
-  ) {
-    this.transpiled = transpile.callable(initializer);
-    this.parameters = this.transpiled.parameters.map(
-      (p) => new Parameter(this.transpile, p)
-    );
-  }
-
-  public render(): Markdown {
-    const md = new Markdown({
-      id: `${this.transpiled.parentType.fqn}.Initializer`,
-      header: {
-        title: "Initializer",
-      },
-    });
-
-    md.code(
-      this.transpile.language,
-      `${this.transpiled.import}`,
-      "",
-      `${this.transpiled.invocation}`
-    );
-
-    for (const parameter of this.parameters) {
-      md.section(parameter.render());
-    }
-
-    return md;
-  }
-}
