+++ src/api/package/assembly.ts
@@ -1,0 +1,20 @@
+import * as spec from "@jsii/spec";
+import { API_PATHS } from "../../constants/url";
+import { getAssetsPath } from "./util";
+
+export const fetchAssembly = async (
+  name: string,
+  version: string,
+  scope?: string
+): Promise<spec.Assembly> => {
+  const assemblyPath = `${getAssetsPath(name, version, scope)}${
+    API_PATHS.ASSEMBLY_SUFFIX
+  }`;
+  const response = await fetch(assemblyPath);
+  if (!response.ok) {
+    throw new Error(
+      `Failed fetching assembly for ${assemblyPath}: ${response.statusText}`
+    );
+  }
+  return response.json();
+};
