+++ src/hooks/useLanguage/useLanguage.test.ts
@@ -49,7 +49,7 @@ describe("useLanguage", () => {
     });
     const { result } = testRender();
 
-    expect(result.current[0]).toEqual("ts"); // The default language
+    expect(result.current[0]).toEqual("typescript"); // The default language
   });
 
   it("checks localStorage for valid language if no url param value", () => {
@@ -63,13 +63,13 @@ describe("useLanguage", () => {
     getItem.mockReturnValueOnce("ruby");
     const { result } = testRender();
 
-    expect(result.current[0]).toEqual("ts");
+    expect(result.current[0]).toEqual("typescript");
   });
 
   it("falls back to default lang if not localStorage or url param", () => {
     getItem.mockReturnValueOnce(null);
     const { result } = testRender();
 
-    expect(result.current[0]).toEqual("ts");
+    expect(result.current[0]).toEqual("typescript");
   });
 });
